backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
ezdxf==1.4.3
opencv-python==4.10.0.84
numpy==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.1
pymupdf==1.24.14
